common utils file + cleanup
