+++ packages/renderer/src/model/QueryMultiWrapper.ts
@@ -1,6 +1,7 @@
 // Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
+import * as Utils from '@ren/utils/CommonUtils';
 import { APIsController } from '@ren/controller/APIsController';
 import { Callbacks } from '@app/callbacks';
 import { MainDebug } from '@ren/utils/DebugUtils';
@@ -12,7 +13,6 @@ import type {
   QueryMultiEntry,
   ApiCallEntry,
 } from '@polkadot-live/types/subscriptions';
-import * as ApiUtils from '@ren/utils/ApiUtils';
 
 const debug = MainDebug.extend('QueryMultiWrapper');
 
@@ -503,9 +503,7 @@ export class QueryMultiWrapper {
               }
 
               // Otherwise, check if action args are the same.
-              if (
-                ApiUtils.arraysAreEqual(task.actionArgs!, innerT.actionArgs!)
-              ) {
+              if (Utils.arraysAreEqual(task.actionArgs!, innerT.actionArgs!)) {
                 task.dataIndex = innerT.dataIndex;
 
                 dataIndexRegistry.push({
